simple login update
