+++ calimero-login/ui/package.json
@@ -12,7 +12,7 @@
     "@headlessui/react": "^1.6.6",
     "@heroicons/react": "^1.0.6",
     "bs58": "^5.0.0",
-    "calimero-auth-sdk": "^0.7.2",
+    "calimero-sdk": "^0.0.1",
     "near-api-js": "^0.45.1",
     "next": "12.2.5",
     "react": "18.2.0",
